move goto_check_java.h/.cpp to the Java frontend
This code is specific to the Java frontend, and this commit moves it there.
