+++ jbmc/src/janalyzer/janalyzer_parse_options.h
@@ -108,8 +108,7 @@ Author: Daniel Kroening, kroening@kroening.com
 #include <goto-programs/show_goto_functions.h>
 #include <goto-programs/show_properties.h>
 
-#include <analyses/goto_check_java.h>
-
+#include <java_bytecode/goto_check_java.h>
 #include <java_bytecode/java_bytecode_language.h>
 
 class abstract_goto_modelt;
