+++ jbmc/src/java_bytecode/Makefile
@@ -9,6 +9,7 @@ SRC = assignments_from_json.cpp \
       expr2java.cpp \
       generic_parameter_specialization_map.cpp \
       generic_parameter_specialization_map_keys.cpp \
+      goto_check_java.cpp \
       jar_file.cpp \
       jar_pool.cpp \
       java_bmc_util.cpp \
