+++ src/analyses/Makefile
@@ -14,7 +14,6 @@ SRC = ai.cpp \
       global_may_alias.cpp \
       goto_check.cpp \
       goto_check_c.cpp \
-      goto_check_java.cpp \
       goto_rw.cpp \
       guard_bdd.cpp \
       guard_expr.cpp \
